+++ config/routes.rb
@@ -1,5 +1,12 @@
 Randomization::Application.routes.draw do
-  resources :projects
+
+  resources :projects do
+    collection do
+      post :add_treatment_arm
+      post :add_stratification_factor
+      post :add_option
+    end
+  end
 
   devise_for :users, controllers: { registrations: 'contour/registrations', sessions: 'contour/sessions', passwords: 'contour/passwords', confirmations: 'contour/confirmations', unlocks: 'contour/unlocks' }, path_names: { sign_up: 'register', sign_in: 'login' }