+++ db/migrate/20130821194213_add_config_to_projects.rb
@@ -1,0 +1,5 @@
+class AddConfigToProjects < ActiveRecord::Migration
+  def change
+    add_column :projects, :config, :text
+  end
+end