+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## 0.1.0
 
+### Enhancements
+- **Project Changes**
+  - Treatment Arms can now be specified along with a weight allocation
+  - Stratification Factors with options can now be specified
+  - Lists are now generated as a product of the stratification factor options
+
 ## 0.0.0 (August 21, 2013)
 
 - Skeleton files to initialize Rails application with testing framework and continuous integration